+++ CHANGES.rst
@@ -17,6 +17,7 @@ Unreleased
 **Deprecated**
 
 - Ability to use :class:`.CommentForest` as an asynchronous iterator.
+- :meth:`.CommentForest.list` no longer needs to be awaited.
 
 **Fixed**